+++ packages/core/CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Update `AbstractVideoIngestionAppliance` to filter corrupt payloads.
 
+### Changed
+- Update `@tvkitchen/base-classes` to version `2.0.0-alpha.1`.
+- Update `@tvkitchen/base-interfaces` to version `4.0.0-alpha.4`.
+- Update `AbstractVideoIngestionAppliance` to populate Payload `origin` instead of `timestamp`.
+- Remove arrow function method definitions from `AbstractVideoIngestionAppliance` and `AbstractAppliance`.
+
 ## [0.6.1] - 2020-04-07
 ### Added
 - `AbstractVideoIngestionAppliance` now generates payloads with timestamps, and accepts an `origin` setting on appliance construction.